refactor: move layout components to layout/ directory
Move navbar, footer, upload-indicator, and features-prefetcher from
src/components/ root into src/components/layout/ subdirectory.

- Convert relative UI imports in navbar.tsx to absolute @/components/ui/ paths
- Update imports in src/app/(main)/layout.tsx
