+++ src/components/layout/upload-indicator.tsx
